demos/writing_tool/pyproject.toml
[project]
name = "writing-tool"
version = "0.1.0"
description = "A tool for writing and editing text with AI assistance."
readme = "README.md"
authors = [{ name = "niits", email = "trungtdxz@gmail.com" }]
requires-python = ">=3.13"

dependencies = [
  "accelerate>=1.9.0",
  "diffusers>=0.34.0",
  "matplotlib>=3.10.3",
  "nbformat>=4.2.0",
  "opencv-python>=4.12.0.88",
  "pillow>=11.3.0",
  "pip>=25.1.1",
  "plotly>=6.2.0",
  "python-dotenv>=1.1.1",
  "requests>=2.32.4",
  "torch>=2.7.1",
  "torchvision>=0.22.1",
  "transformers>=4.54.1",
]

[dependency-groups]
dev = [
  "black[jupyter]>=25.1.0",
  "isort>=6.0.1",
  "nvitop>=1.5.2",
  "pylint>=3.3.7",
]
# PyTorch index configurations for different CUDA versions and CPU
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
# Default to CUDA 12.8, but can be overridden with environment variables or manual selection
torch = [
  # CPU version (uncomment to use CPU-only PyTorch)
  # { index = "pytorch-cpu" },

  # CUDA versions (uncomment the desired version)
  # { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  # { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  # { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  # CPU version (uncomment to use CPU-only torchvision)
  # { index = "pytorch-cpu" },

  # CUDA versions (uncomment the desired version)
  # { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  # { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  # { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[project.scripts]
writing-tool = "writing_tool:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
